Move more into stats file
