+++ driver/driver.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"strings"
 	"time"
 
 	"github.com/hashicorp/go-hclog"
@@ -14,7 +13,6 @@ import (
 	"github.com/hashicorp/nomad/plugins/base"
 	"github.com/hashicorp/nomad/plugins/drivers"
 	"github.com/hashicorp/nomad/plugins/shared/hclspec"
-	"github.com/shirou/gopsutil/v3/process"
 )
 
 const (
@@ -506,47 +504,4 @@ func (d *Driver) ExecTaskStreaming(ctx context.Context, taskID string, opts *dri
 
 func (d *Driver) generateVMName(allocationID string) string {
 	return fmt.Sprintf("nomad-%s", allocationID)
-}
-
-// relatedPIDs returns the tart process PID and any virtualization service PIDs
-// associated with the given VM name.
-func (d *Driver) relatedPIDs(vmName string, tartPID int) []int {
-	pids := []int{}
-	if tartPID > 0 {
-		pids = append(pids, tartPID)
-	}
-
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return pids
-	}
-
-	diskPath := filepath.Join(home, ".tart", vmName, "disk.img")
-
-	processes, err := process.Processes()
-	if err != nil {
-		return pids
-	}
-
-	for _, p := range processes {
-		exe, err := p.Exe()
-		if err != nil {
-			continue
-		}
-		if !strings.Contains(exe, "Virtualization.VirtualMachine") {
-			continue
-		}
-
-		files, err := p.OpenFiles()
-		if err != nil {
-			continue
-		}
-		for _, f := range files {
-			if f.Path == diskPath {
-				pids = append(pids, int(p.Pid))
-				break
-			}
-		}
-	}
-	return pids
 }
